requirements.txt
fastapi
langchain
langchain_community

langchain_core

langchain_google_genai
langchain_groq

langserve
python-dotenv
Requests

streamlit

uvicorn

